# SCIM: Optimize user queries: Use WITH and JOIN rather than subqueries