--- a/internal/database/users.go
+++ b/internal/database/users.go
@@ -1192,6 +1192,7 @@ func (u *userStore) getOneBySQL(ctx context.Context, q *sqlf.Query) (*types.User
 // getBySQL returns users matching the SQL query, if any exist.
 func (u *userStore) getBySQL(ctx context.Context, query *sqlf.Query) ([]*types.User, error) {
 	q := sqlf.Sprintf(`
+WITH scim_accounts AS (SELECT user_id FROM user_external_accounts WHERE service_type = 'scim')
 SELECT u.id,
        u.username,
        u.display_name,
@@ -1204,8 +1205,10 @@ SELECT u.id,
        u.invalidated_sessions_at,
        u.tos_accepted,
        u.searchable,
-       (SELECT COUNT(user_id) FROM user_external_accounts WHERE user_id=u.id AND service_type = 'scim') >= 1 AS scim_controlled
-FROM users u %s`, query)
+       (SELECT COUNT(user_id) FROM scim_accounts WHERE user_id=u.id) >= 1 AS scim_controlled
+FROM users u
+LEFT JOIN scim_accounts sa ON u.id = sa.user_id
+ %s`, query)
 	rows, err := u.Query(ctx, q)
 	if err != nil {
 		return nil, err
@@ -1232,6 +1235,14 @@ FROM users u %s`, query)
 }
 
 const userForSCIMQueryFmtStr = `
+WITH scim_accounts AS (
+    SELECT
+        user_id,
+        account_id,
+        account_data
+    FROM user_external_accounts
+    WHERE service_type = 'scim'
+)
 SELECT u.id,
        u.username,
        u.display_name,
@@ -1245,9 +1256,11 @@ SELECT u.id,
        u.tos_accepted,
        u.searchable,
        ARRAY(SELECT email FROM user_emails WHERE user_id = u.id AND verified_at IS NOT NULL) AS emails,
-       (SELECT account_id FROM user_external_accounts WHERE user_id=u.id AND service_type = 'scim') AS scim_external_id,
-       (SELECT account_data FROM user_external_accounts WHERE user_id=u.id AND service_type = 'scim') AS scim_account_data
-  FROM users u %s`
+       sa.account_id AS scim_external_id,
+       sa.account_data AS scim_account_data
+FROM users u
+LEFT JOIN scim_accounts sa ON u.id = sa.user_id
+%s`
 
 // getBySQLForSCIM returns users matching the SQL query, along with their email addresses and SCIM ExternalID.
 func (u *userStore) getBySQLForSCIM(ctx context.Context, query *sqlf.Query) ([]*types.UserForSCIM, error) {