@vdavid 2023-03-08
Thank you for all your contributions, everyone! I went with @eseliger's solution for the first query and incorporated @pjlast's thought about filtering for non-deleted external accounts. Then I ran the `EXPLAIN ANALYZE` command a few more times to ensure that the final state of both queries is performant (it is). Now I'm confident to merge this.